\name{pred.mclust}
\alias{pred.mclust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Cluster Metrics
}
\description{
Used to evaluate metrics of classified data which uses clustering based on EM for GMM using Mclust package.
}
\usage{
pred.mclust(testdata, traindata)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{testdata}{
A data frame used for testing a model. (confirmatory data set)
}
  \item{traindata}{
A data frame used for training a model.
}
}
\details{
A function that assigns cluster to each test data point using clusters generated by mclust package and Then detects whether the data point is a male or female and evaluates the performance of the model using 3 metrics namely Accuracy, False Positive and False Negative rates.
}
\value{
 \item{Accuracy }{Total correcly classified data points among all the test data.}
 \item{False Positive Rate }{When it's actually no, how often does it predict yes}
  \item{False Negative Rate }{When it's actually Yes, how often does it predict no}

}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Sai Kiran Vamaraju
}
\note{
Only works on a specific dataset. Not Applicable for all kinds of data.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (testdata, traindata)
{
    mc = Mclust(traindata[, (1:3)], G = 2)
    tc <- predict.Mclust(mc, testdata[, 1:3])
    clustert <- data.frame(testdata, tc)
    clustert$classification[clustert$classification == 1] <- "Female"
    clustert$classification[clustert$classification == 2] <- "Male"
    cmo <- confusionMatrix(clustert$classification, clustert$Gender)
    cm <- cmo$table
    acc <- sum(diag(cm))/sum(cm)
    fp <- cm[1, 2]/sum(cm[, 2])
    fn <- cm[2, 1]/sum(cm[, 1])
    ls <- list(Accuracy = acc, `False Positive` = fp, `False Negative` = fn)
    return(ls)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
